Make content parameter Sensitive
Right now Puppet has no ability to customize the redaction value of a string, which would be required to produce a partially redacted diff from a file. This change demonstrates the closest thing we can do today, which is let Puppet know that the entire value of the file is Sensitive. That will protect the file content string in reports, etc.

File diff output on the command line and in logs may be different, special. I don't know if show_diff understands or is able to account for file resources including Sensitive content.